backend/pyproject.toml
[tool.poetry]
name = "mis-eventos-backend"
version = "0.1.0"
description = "Backend MVP for Mis Eventos"
authors = ["Carlos Garzon"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.6" }
sqlmodel = "^0.0.22"
psycopg = { extras = ["binary"], version = "^3.2.1" }
alembic = "^1.13.2"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic-settings = "^2.5.2"
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.2"
ruff = "^0.6.9"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
